Create makefile to manage repo
The overall goal of the makefile is to document clearly with
useful and working code, the necessary commands for managing
the repo.  This includes changelog updates, setting up git
hooks, creating the requirements file.

Later we'll add commands for updating the version number, creating
and deploying a package, checking pylint score, and so forth.