Pass exchange test
There is a discrepancy in how the python pool code was
handling exchanges for metapools versus basepools.  In
particular, whether the amounts input and output were
in D units or native units.

Code was updated so that `exchange` follows the smart
contract logic, i.e. inputs and outputs use native
token units.