Fix `arberror` to match changes to `exchange`
Prior changes to `exchange` ensure adherence to the smart contract
method's assumption that `dx` is in native token units (not units
of D).  This is *partially* at odds with the usage in the sim code,
which sometimes uses `exchange` assuming that `dx` is in D units.

This assumption by the sim code doesn't actual hold in the codepaths
where the pool is a metapool and the primary stable is being swapped
with a basepool underlyer.  In those cases, `dx` is either multiplied
by rates or used as a token amount to add liquidity.

Thus the changes not only adhere to the smart contract conventions
but also makes `exchange` internally consistent.

However, the issue here is that the sim code generates trades,
assuming D units for `dx`.  The bounds for the root solver, in
particular, are set using D units.  This doesn't seem to actually
pose an issue when testing because when the trades are applied to
the pool, the same logic is used again, so generates the right
results.

Potentially this could be problematic though, since the lower bound
is higher than expected (10**12 rather than 1, for a 6 decimals coin).
Thus we fix the `arberror` logic to convert `dx` to native token units
to prevent potential issues.