Remove deprecated/unused pool methods
These are no longer used in the sim code.  We will want some still
to adhere to the smart contract interfaces, but for that use-case,
we will want to revamp the implementations to match anyway.

For now, I'm removing to declutter the codebase and we will work
these in methodically with test coverage.