Setup linting commands
Makefile now has a lint "pipeline" command, which will go through
`black`, `flake8`, and `pylint`.  The pylint process uses a script
to check the score against a threshold.  It will fail if there is a
serious pylint error (as distinguished from a "warning") and/or
the score is below threshold.

Currently threshold is set to 9.4 which is a little below the repo's
score after all the cleanups.