Rework __init__ logic
Our main use-case is when balances are provided.  So this is made
explicit now, instead of relying upon the crufty D interface
from the old curve repo code.

When balances are provided, D can be optionally provided.  This
helps avoid to calculate D and simplifies testing.

When balances are not provided, D must be provided.  Then balances
are derived from that, assuming the pool is perfectly balanced.

Virtual price and tokens is set appropriately.